Fixed problem with flushAll call when in read only mode. Fixed problem with log marked a clean and keepLogicalLogs=false trying to delete the log file twice.
git-svn-id: https://svn.neo4j.org/components/kernel/trunk@4481 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ enterprise/src/main/java/org/neo4j/kernel/impl/nioneo/xa/NeoStoreXaDataSource.java
@@ -216,7 +216,10 @@ public class NeoStoreXaDataSource extends XaDataSource
 
     public void close()
     {
-        neoStore.flushAll();
+        if ( !readOnly )
+        {
+            neoStore.flushAll();
+        }
         xaContainer.close();
         if ( logApplied )
         {
